change slave at float registers

+++ src/main/java/com/vk/repository/modbus/IModbusRepositoryRightWinderPLC110.java
@@ -5,13 +5,13 @@ package com.vk.repository.modbus;
  */
 public interface IModbusRepositoryRightWinderPLC110 {
 
-    public void writeDataToRegister0(final int value);
+    public void writeDataToRegister0(final float value);
     public void readDataFromRegister0(final boolean enableBatch);
-    public void writeDataToRegister1(final int value);
+    public void writeDataToRegister1(final float value);
     public void readDataFromRegister1(final boolean enableBatch);
-    public void writeDataToRegister2(final int value);
+    public void writeDataToRegister2(final float value);
     public void readDataFromRegister2(final boolean enableBatch);
-    public void writeDataToRegister3(final int value);
+    public void writeDataToRegister3(final float value);
     public void readDataFromRegister3(final boolean enableBatch);
     public void writeDataToRegister4(final int value);
     public void readDataFromRegister4(final boolean enableBatch);
